requirements.txt
astroid==2.3.3
autopep8==1.4.4
beautifulsoup4==4.8.1
boilerpipe3==1.1
certifi==2019.9.11
charade==1.0.3
chardet==3.0.4
idna==2.8
isort==4.3.21
JPype1-py3==0.5.5.4
jusText==2.2.0
langid==1.1.6
lazy-object-proxy==1.4.3
lxml==4.4.1
mccabe==0.6.1
numpy==1.17.4
pycodestyle==2.5.0
pylint==2.4.4
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
urllib3==1.25.7
wrapt==1.11.2
